requirements.txt
fastapi
uvicorn
openai
langchain==0.1.16
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
python-dotenv